moved handleSubmit to Header component
